+++ SuperCyanTweaks/Enemies/Geep.cs
@@ -1,7 +1,4 @@
 ﻿using RoR2;
-using RoR2.CharacterAI;
-using R2API;
-using UnityEngine;
 using UnityEngine.AddressableAssets;
 
 namespace SuperCyanTweaks